+++ Source/HigherPower.cs
@@ -8,9 +8,6 @@
 		{
 			MultiModPatch.PoweredGenerators.Add(GetDatabaseThing("ChemfuelPoweredGenerator_II"));
 			MultiModPatch.SolarGenerators.Add(GetDatabaseThing("SolarGenerator_II"));
-			MultiModPatch.GeothermalGenerators.Add(GetDatabaseThing("GeothermalGenerator_II"));
-			MultiModPatch.WindTurbines.Add(GetDatabaseThing("WindTurbine_II"));
-			MultiModPatch.WatermillGenerators.Add(GetDatabaseThing("WatermillGenerator_II"));
 		}
 	}
 }