+++ Source/MultiModPatch.cs
@@ -14,6 +14,14 @@ namespace Replace_Stuff_Compatibility
 		public static List<ThingDef> Smithys = new List<ThingDef>()
 			{GetDatabaseThing("FueledSmithy"), GetDatabaseThing("ElectricSmithy")};
 
+		public static List<ThingDef> MachiningTables = new List<ThingDef>() {GetDatabaseThing("TableMachining")};
+
+		public static List<ThingDef> Stoves = new List<ThingDef>()
+			{GetDatabaseThing("FueledStove"), GetDatabaseThing("ElectricStove")};
+
+		public static List<ThingDef> TailoringBenches = new List<ThingDef>()
+			{GetDatabaseThing("HandTailoringBench"), GetDatabaseThing("ElectricTailoringBench")};
+
 		public static List<ThingDef> Lights = new List<ThingDef>() { };
 
 		public static List<ThingDef> Sunlamps = new List<ThingDef>() {GetDatabaseThing("SunLamp")};
@@ -26,12 +34,6 @@ namespace Replace_Stuff_Compatibility
 			{GetDatabaseThing("WoodFiredGenerator"), GetDatabaseThing("ChemfuelPoweredGenerator")};
 
 		public static List<ThingDef> SolarGenerators = new List<ThingDef>() {GetDatabaseThing("SolarGenerator")};
-
-		public static List<ThingDef> GeothermalGenerators = new List<ThingDef>() {GetDatabaseThing("GeothermalGenerator")};
-
-		public static List<ThingDef> WindTurbines = new List<ThingDef>() {GetDatabaseThing("WindTurbine")};
-
-		public static List<ThingDef> WatermillGenerators = new List<ThingDef>() {GetDatabaseThing("WatermillGenerator")};
 		
 		public static List<ThingDef> TVs = new List<ThingDef>()
 		{
@@ -39,10 +41,14 @@ namespace Replace_Stuff_Compatibility
 			GetDatabaseThing("MegascreenTelevision")
 		};
 
-		public static List<ThingDef> Fabricators = new List<ThingDef>() { GetDatabaseThing("FabricationBench") };
+		public static List<ThingDef> Fabricators = new List<ThingDef>() {GetDatabaseThing("FabricationBench")};
 
 		public static List<ThingDef> Wardrobes = new List<ThingDef>() { };
-		
+
+		public static List<ThingDef> ArtTables = new List<ThingDef>() { GetDatabaseThing("TableSculpting") };
+
+		public static List<ThingDef> Columns = new List<ThingDef>() { GetDatabaseThing("Column") };
+
 		protected override void AddItems()
 		{
 			// Allow all "plant growable items" to replace each other, and when they do attempt to set the growing plant type
@@ -56,6 +62,18 @@ namespace Replace_Stuff_Compatibility
 			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(
 				building => typeof(Building_Battery).IsAssignableFrom(building.thingClass)));
 
+			// We can use placeWorkers and comps to check what kind of power is being generated so that we don't have to worry
+			// about each item individually
+			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(
+				building => building.placeWorkers?.Any(placeWorker =>
+					placeWorker == typeof(PlaceWorker_WatermillGenerator)) ?? false));
+			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(
+				building => building.placeWorkers?.Any(placeWorker =>
+					placeWorker == typeof(PlaceWorker_WindTurbine)) ?? false));
+			NewThingReplacement.replacements.Add(new NewThingReplacement.Replacement(
+				building => building.placeWorkers?.Any(placeWorker =>
+					placeWorker == typeof(PlaceWorker_OnSteamGeyser)) ?? false));
+
 			AddInterchangeableWorkbenches(Smelters);
 			AddInterchangeableWorkbenches(Smithys);
 			AddInterchangeableList(Lights);
@@ -65,11 +83,13 @@ namespace Replace_Stuff_Compatibility
 			AddInterchangeableList(TVs);
 			AddInterchangeableList(PoweredGenerators);
 			AddInterchangeableList(SolarGenerators);
-			AddInterchangeableList(GeothermalGenerators);
-			AddInterchangeableList(WindTurbines);
-			AddInterchangeableList(WatermillGenerators);
 			AddInterchangeableList(Fabricators);
 			AddInterchangeableList(Wardrobes);
+			AddInterchangeableList(TailoringBenches);
+			AddInterchangeableList(Stoves);
+			AddInterchangeableList(MachiningTables);
+			AddInterchangeableList(ArtTables);
+			AddInterchangeableList(Columns);
 		}
 	}
 }